requirements.txt
uvicorn~=0.29.0
fastapi~=0.110.0
SQLAlchemy~=2.0.28
aiosqlite~=0.20.0
pydantic~=2.7.1
pydantic-settings~=2.2.1
PyJWT~=2.8.0
passlib~=1.7.4
bcrypt==4.0.1
python-dotenv~=1.0.1
aiohttp~=3.9.5
cryptography==42.0.7
email_validator==2.1.1
fastapi-mail==1.4.1
python-multipart==0.0.9
redis~=5.0.4
asyncpg==0.29.0
alembic~=1.13.1